Move CNI commands from cmd/main to package/plugin
Move implementation of CNI handlers to package/plugin
